backend/requirements.txt
uvicorn
python-dotenv
fastapi
starlette
requests
sqlalchemy
pydantic